LPS-174541 Move to the right package
